Trying alternate lookup for '0 snaps' label.
Android emulator doesn't actually have back button, so trying alternate
methods for that. Might make android branch if screen is always a known
size, but didn't have time today.

Cleaned up error messages if buttons can't be found to be less
confusing, and rewrote the board printer to use letters and one-based
indices for grid locations to make it easier for non-developers to
understand.